fix validate issues
