+++ components/Product/ValidateFields.tsx
