+++ package.json
@@ -21,6 +21,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-native": "^0.66.4",
+    "redux-form": "^8.3.8",
     "swr": "^1.0.1"
   },
   "devDependencies": {
